[test] reset miner mocktime in maxtipage to prevent time offset warning
This change is necessary for the next commit, where the logic used to warn
the user if their clock seems out of sync with the network is changed.

By resetting the node_miner time to its actual system time at the end of
each test, we ensure there is no time offset with node_ibd the next time
the function is called, ensuring TimeOffsets::WarnIfOutOfSync() doesn't
output to stderr, failing the test.